Compatibility with multi column, and hooks
Compatibility between this add-on and multi column window has been
broken for some time. I did repair it in an add-on merging them, but
I'm not a fan of this solution.

Instead, here, I decided that instead of redefining entirely loadNote,
I wrap code around it. The code just add the frozen image.

In js, setFrozenFields is just in charge of adding the image before
the name. The image is .9em in height, so smaller than the text. It
even leads to more space in the field